+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Map2: fixed empty content processing of My Locations.
 - Fixed Food Filters validation in Food Filters settings page [#5502](https://github.com/rokwire/illinois-app/issues/5502).
+- Browse Panel does not modify FlexUI's content any more [#5502](https://github.com/rokwire/illinois-app/issues/5502).
 
 ## [7.3.17] - 2025-10-22
 ### Changed